requirements.txt
Flask
mysql-connector-python
Flask-Bcrypt
flask_cors